requirements.txt
snap-stanford==6.0.0
terminaltables==3.1.0
